# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Book My Cleaning — AI receptionist for Jobber cleaning companies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /company/quo/connect:
    post:
      operationId: connectQuo
      tags: [phone]
      summary: Connect this company's own Quo workspace with their API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectQuoInput"
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
  /quo/numbers:
    get:
      operationId: listQuoNumbers
      tags: [phone]
      summary: List the real phone numbers in the connected Quo workspace
      responses:
        "200":
          description: Quo numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoNumber"
    post:
      operationId: selectQuoNumbers
      tags: [phone]
      summary: Choose which Quo lines the receptionist watches and register webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoNumberSelection"
      responses:
        "200":
          description: Watched lines updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
components:
  schemas:
...
    CustomQuestion:
      type: object
      required: [question, answer]
      properties:
        question: { type: string }
        answer: { type: string }
...
    ConnectQuoInput:
      type: object
      required: [apiKey]
      properties:
        apiKey: { type: string }
    QuoNumberSelection:
      type: object
      required: [numberIds]
      properties:
        numberIds:
          type: array
          items: { type: string }
    QuoNumber:
      type: object
      required: [id, phoneNumber]
      properties:
        id: { type: string }
        phoneNumber: { type: string }
        name: { type: ["string", "null"] }
        watched: { type: boolean }
    WatchedNumber:
      type: object
      required: [id, phoneNumber, name, watched]
      properties:
        id: { type: string }
        phoneNumber: { type: string }
        name: { type: string }
        watched: { type: boolean }
    SetupStatus:
      type: object
      required: [accountCreated, jobberConnected, quoConnected, phoneProvisioned, receptionistConfigured, teamInvited, isLive, completedSteps, totalSteps]
      properties:
        accountCreated: { type: boolean }
        jobberConnected: { type: boolean }
        quoConnected: { type: boolean }
        phoneProvisioned: { type: boolean }
        receptionistConfigured: { type: boolean }
        teamInvited: { type: boolean }
        isLive: { type: boolean }
        completedSteps: { type: integer }
        totalSteps: { type: integer }
    Company:
      type: object
      required: [id, name, greeting, collectFields, customQuestions, jobberConnected, quoConnected, watchedNumbers, isLive, setupStatus, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        greeting: { type: string }
        collectFields:
          type: array
          items: { type: string }
        customQuestions:
          type: array
          items:
            $ref: "#/components/schemas/CustomQuestion"
        ringThroughNumber: { type: ["string", "null"] }
        phoneNumber: { type: ["string", "null"] }
        jobberConnected: { type: boolean }
        jobberAccountName: { type: ["string", "null"] }
        quoConnected: { type: boolean }
        quoWorkspaceName: { type: ["string", "null"] }
        quoKeyLast4: { type: ["string", "null"] }
        watchedNumbers:
          type: array
          items:
            $ref: "#/components/schemas/WatchedNumber"
        isLive: { type: boolean }
        setupStatus:
          $ref: "#/components/schemas/SetupStatus"
        createdAt: { type: string }
...